% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_time_loading}
\alias{plot_time_loading}
\title{Plot the temporal loading functions (multi-modality)}
\usage{
plot_time_loading(res, r = NULL, ...)
}
\arguments{
\item{res}{Output of \code{\link{multi_tempted_decomp}} or
\code{\link{multitempted_all}}.}

\item{r}{Number of components to plot. Default: all components in \code{res}.}

\item{...}{Additional aesthetics passed to \code{ggplot2::geom_line()}.}
}
\value{
A ggplot2 object.
}
\description{
Plots the temporal loading functions (\code{Zeta_hat}) estimated
by \code{\link{multi_tempted_decomp}}, faceted by modality and coloured by
component.
}
\seealso{
\code{\link{multitempted_all}}, \code{\link{multi_tempted_decomp}}.
}
